+++ src/quickjs_utils/mod.rs
@@ -7,6 +7,7 @@ pub mod compile;
 pub mod dates;
 pub mod errors;
 pub mod functions;
+pub mod interrupthandler;
 pub mod iterators;
 pub mod json;
 pub mod maps;