+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.4.2 (work in progress)
 
-* 
+* moved reflection code to reflection/mod.rs (should not affect api)
+* toPrimitive for Proxy classes (do stuff like console.log('got: ' + MyProxyInstanceOrClass))
 
 # 0.4.1